pyproject.toml
[project]
name = "ai-skill-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imbalanced-learn>=0.13.0",
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "markitdown[all]>=0.1.5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "onnxruntime>=1.20.1",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "xgboost>=3.2.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
